+++ src/code-actions-provider/index.ts
@@ -1,5 +1,5 @@
 import * as vscode from 'vscode';
-import { DART_SELECTOR } from '../utils/document-selector';
+import { DART_SELECTOR } from '../utils/document.selector';
 import { UtilityActionProvider } from './utility-code-action-provider';
 
 export function activateCodeActionProvider(context: vscode.ExtensionContext) {