+++ src/code-actions-provider/utility-action-type.ts
@@ -1,5 +1,6 @@
 export enum UtilityActionType {
   generateToString,
+  generateConstructor,
   generateCopyWith,
   generateEquatable,
   implementJsonSerializable,