+++ src/completion-provider/index.ts
@@ -1,5 +1,5 @@
 import * as vscode from 'vscode';
-import { DART_SELECTOR } from '../utils/document-selector';
+import { DART_SELECTOR } from '../utils/document.selector';
 import DartCompletionProvider from './dart-completion-provider';
 
 export function activateCompletionProvider(context: vscode.ExtensionContext) {