+++ src/providers/index.ts
@@ -1,6 +1,6 @@
 import * as vscode from 'vscode';
-import { ConfigurationUtils } from '../utils/configuration-utils';
-import { DART_SELECTOR } from '../utils/document-selector';
+import { ConfigurationUtils } from '../utils/configuration.utils';
+import { DART_SELECTOR } from '../utils/document.selector';
 import { TestLensProvider } from './test-lens-provider';
 
 export function activateProviders(context: vscode.ExtensionContext) {