+++ src/utils/type.utils.ts
@@ -1,0 +1,3 @@
+export function isTypeNullable(type: string) {
+  return type.endsWith('?');
+}